ref: move `GameWorld` to `World` namespace
