+++ src/CoreLib/Game/GameServer.cs
@@ -19,6 +19,7 @@ using Imlight.CoreLib.WizardData.Implementations;
 using Imlight.CoreLib.Game.Commands;
 using Imlight.CoreLib.WizardData.Databases;
 using Imlight.CoreLib.Game.Effects;
+using Imlight.CoreLib.Game.World;
 
 namespace Imlight.CoreLib.Game;
